Collecting Last Update of Project

+++ AdminDashboardB/Controllers/OrderController.cs
@@ -1,11 +1,8 @@
 ﻿using ApplicationB.Services_B.Order;
-using DTOsB.OrderDTO;
-using DTOsB.Shared;
+using DTOsB.Order.OrderDTO;
 using Microsoft.AspNetCore.Mvc;
-using ModelsB.Authentication_and_Authorization_B;
-using ModelsB.Order_B;
 
-namespace AdminDashboardB.Controllers
+namespace DTOsB.Controllers
 {
     public class OrderController : Controller
     {
@@ -27,8 +24,8 @@ namespace AdminDashboardB.Controllers
         public async Task<IActionResult> Add()
         {
             AddOrUpdateOrderBDTO order = new AddOrUpdateOrderBDTO();
-           // order.Users = 
-          
+            // order.Users = 
+
             return View("Add", order);
         }
 
@@ -37,12 +34,13 @@ namespace AdminDashboardB.Controllers
         {
             /* (ModelState.IsValid)
             {*/
-                orderBDTO.OrderDate = DateTime.Now;
-                orderBDTO.TotalPrice = 5000;
-                orderBDTO.ApplicationUserId = "db0a8336-7f0f-416c-90c8-a8dfd01d97f7";
+            orderBDTO.OrderDate = DateTime.Now;
+            orderBDTO.TotalPrice = 5000;
+            orderBDTO.ApplicationUserId = "db0a8336-7f0f-416c-90c8-a8dfd01d97f7";
 
-                await orderService.CreateOrderAsync(orderBDTO);
-                return RedirectToAction("Index");
+
+            await orderService.CreateOrderAsync(orderBDTO);
+            return RedirectToAction("Index");
             /*}
 
             return View(orderBDTO);*/
@@ -53,11 +51,12 @@ namespace AdminDashboardB.Controllers
         public async Task<IActionResult> Edit(int id)
         {
             SelectOrderBDTO order = await orderService.GetOrderByIdAsync(id);
+
             if (order == null || order.CurrentStatus == ModelsB.Order_B.Status.Shipped || order.CurrentStatus == ModelsB.Order_B.Status.Delivered)
             {
-                return BadRequest("Cannot edit this order.");
+                TempData["ErrorMessage"] = "You can't edit this order anymore.";
+                return RedirectToAction("Index");
             }
-
             var updateOrder = new AddOrUpdateOrderBDTO()
             {
                 Id = id,
@@ -66,7 +65,8 @@ namespace AdminDashboardB.Controllers
                 CurrentStatus = order.CurrentStatus,
                 ApplicationUserId = "db0a8336-7f0f-416c-90c8-a8dfd01d97f7"
             };
-            return View("Edit", updateOrder);  
+
+            return View("Edit", updateOrder);
         }
 
         [HttpPost]
@@ -78,7 +78,7 @@ namespace AdminDashboardB.Controllers
                 return RedirectToAction("Index");
             }
 
-            return View(orderBDTO); 
+            return View(orderBDTO);
         }
 
         //==============Details=====================
@@ -94,37 +94,14 @@ namespace AdminDashboardB.Controllers
         public async Task<IActionResult> CancelOrder(int id)
         {
             var order = await orderService.GetOrderByIdAsync(id);
-            if (order == null)
+            if (order == null || order.CurrentStatus == ModelsB.Order_B.Status.Shipped || order.CurrentStatus == ModelsB.Order_B.Status.Delivered)
             {
-                return NotFound();
+                TempData["ErrorMessage"] = "You can't cancel this order anymore.";
+                return RedirectToAction("Index");
             }
 
-            await orderService.DeleteOrderAsync(id); 
+            await orderService.DeleteOrderAsync(id);
             return RedirectToAction(nameof(Index));
-        }
-
-
-
-
-
-
-
-
-
-
-
-        public async Task<IActionResult> UpdateOrderDetails()
-        {
-            return RedirectToAction();
-        }
-        public async Task<IActionResult> StartProccess()
-        {
-            return RedirectToAction();
-        }
-
-        public async Task<IActionResult> StartShip()
-        {
-            return RedirectToAction();
         }
     }
 }
